Refactored to make registration and unregistration more symmetric.
Added a submit(Runnable) method to Selector, to allow to interact
with the Selector, and moved registration and unregistration tasks
to the Channel implementation (from the Selector implementation).
Made registration and unregistration synchronous, so that no race
conditions apply between the two operations (e.g. closing a channel
before it's registered).