Removed the prepare event.
Turns out that this event is not needed; it was used to allow the
blocking connection to remain with a disabled read interest, but
working with SSL showed that it is not possible.
In the SSL case, the other peer can write immediately and this
data will come down in the same TCP packet as the SSL handshake.
So the read will be notified in any case to the blocking connection.
In light of this, the connection will always be with read interest
enabled, and the blocking connection will have to buffer up any
early read (which would have to do in any case with SSL).
Now, only after the open event (to avoid that read events are
dispatched before open events), the read interest is enabled.